# Optimise carousel beatmap filter method