--- a/osu.Game/Beatmaps/BeatmapInfoExtensions.cs
+++ b/osu.Game/Beatmaps/BeatmapInfoExtensions.cs
@@ -1,7 +1,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
+using System.Collections.Generic;
 using osu.Framework.Localisation;
 
 namespace osu.Game.Beatmaps
@@ -29,25 +29,19 @@ namespace osu.Game.Beatmaps
             return new RomanisableString($"{metadata.GetPreferred(true)}".Trim(), $"{metadata.GetPreferred(false)}".Trim());
         }
 
-        public static ReadOnlySpan<string> GetSearchableTerms(this IBeatmapInfo beatmapInfo)
+        public static List<string> GetSearchableTerms(this IBeatmapInfo beatmapInfo)
         {
-            string[] terms = new string[8];
-            int i = 0;
-            var metadata = beatmapInfo.Metadata;
+            var termsList = new List<string>(BeatmapMetadataInfoExtensions.MAX_SEARCHABLE_TERM_COUNT + 1);
+
             addIfNotNull(beatmapInfo.DifficultyName);
-            addIfNotNull(metadata.Author.Username);
-            addIfNotNull(metadata.Artist);
-            addIfNotNull(metadata.ArtistUnicode);
-            addIfNotNull(metadata.Title);
-            addIfNotNull(metadata.TitleUnicode);
-            addIfNotNull(metadata.Source);
-            addIfNotNull(metadata.Tags);
-            return terms.AsSpan(0, i);
+
+            BeatmapMetadataInfoExtensions.CollectSearchableTerms(beatmapInfo.Metadata, termsList);
+            return termsList;
 
             void addIfNotNull(string? s)
             {
                 if (!string.IsNullOrEmpty(s))
-                    terms[i++] = s;
+                    termsList.Add(s);
             }
         }