--- a/osu.Game/Beatmaps/BeatmapMetadataInfoExtensions.cs
+++ b/osu.Game/Beatmaps/BeatmapMetadataInfoExtensions.cs
@@ -3,7 +3,7 @@
 
 #nullable disable
 
-using System.Linq;
+using System.Collections.Generic;
 using osu.Framework.Localisation;
 
 namespace osu.Game.Beatmaps
@@ -13,16 +13,31 @@ namespace osu.Game.Beatmaps
         /// <summary>
         /// An array of all searchable terms provided in contained metadata.
         /// </summary>
-        public static string[] GetSearchableTerms(this IBeatmapMetadataInfo metadataInfo) => new[]
+        public static string[] GetSearchableTerms(this IBeatmapMetadataInfo metadataInfo)
         {
-            metadataInfo.Author.Username,
-            metadataInfo.Artist,
-            metadataInfo.ArtistUnicode,
-            metadataInfo.Title,
-            metadataInfo.TitleUnicode,
-            metadataInfo.Source,
-            metadataInfo.Tags
-        }.Where(s => !string.IsNullOrEmpty(s)).ToArray();
+            var termsList = new List<string>(MAX_SEARCHABLE_TERM_COUNT);
+            CollectSearchableTerms(metadataInfo, termsList);
+            return termsList.ToArray();
+        }
+
+        internal const int MAX_SEARCHABLE_TERM_COUNT = 7;
+
+        internal static void CollectSearchableTerms(IBeatmapMetadataInfo metadataInfo, IList<string> termsList)
+        {
+            addIfNotNull(metadataInfo.Author.Username);
+            addIfNotNull(metadataInfo.Artist);
+            addIfNotNull(metadataInfo.ArtistUnicode);
+            addIfNotNull(metadataInfo.Title);
+            addIfNotNull(metadataInfo.TitleUnicode);
+            addIfNotNull(metadataInfo.Source);
+            addIfNotNull(metadataInfo.Tags);
+
+            void addIfNotNull(string s)
+            {
+                if (!string.IsNullOrEmpty(s))
+                    termsList.Add(s);
+            }
+        }
 
         /// <summary>
         /// A user-presentable display title representing this metadata.