Might help a bit with #11571
With this the total number of memory allocations is reduced to 1 and it takes less time.
I profiled the method and the `GetSearchableTerms` goes from 43.8% to 27.8% of the total time of `Filter`