@bdach 2023-03-03
This PR should include a BenchmarkDotNet benchmark for result reproduction. We have a [special project](https://github.com/ppy/osu/tree/master/osu.Game.Benchmarks) for those.