@peppy 2023-03-04
I'm not against this change (after confirming with benchmarks and depending on review complexity) but do note that the main overhead is the actual retrieval from realm, and fixing that requires moving as much of the query to realm-supported LINQ or realm query language.