@OliBomby 2023-03-04
I created some benchmarks based on the existing unit tests and here are the results:

**Before my changes**
```
|                          Method |        Mean |     Error |    StdDev |      Median |   Gen0 | Allocated |
|-------------------------------- |------------:|----------:|----------:|------------:|-------:|----------:|
|           CarouselBeatmapFilter |    29.62 ns |  0.609 ns |  0.570 ns |    29.30 ns |      - |         - |
| CriteriaMatchingSpecificRuleset |    62.60 ns |  1.233 ns |  1.514 ns |    61.65 ns |      - |         - |
|        CriteriaMatchingRangeMax |    82.59 ns |  1.641 ns |  2.300 ns |    81.35 ns |      - |         - |
|           CriteriaMatchingTerms | 1,437.47 ns | 28.611 ns | 36.184 ns | 1,423.41 ns | 0.1011 |     856 B |
|         CriteriaMatchingCreator |    97.00 ns |  1.213 ns |  0.947 ns |    97.25 ns |      - |         - |
|      CriteriaMatchingBeatmapIDs | 1,149.67 ns | 21.988 ns | 24.440 ns | 1,133.94 ns | 0.0877 |     736 B |
```
**After my changes**
```
|                          Method |      Mean |     Error |    StdDev |    Median |   Gen0 | Allocated |
|-------------------------------- |----------:|----------:|----------:|----------:|-------:|----------:|
|           CarouselBeatmapFilter |  32.21 ns |  0.670 ns |  1.617 ns |  31.60 ns |      - |         - |
| CriteriaMatchingSpecificRuleset |  64.65 ns |  1.115 ns |  1.239 ns |  64.28 ns |      - |         - |
|        CriteriaMatchingRangeMax |  84.45 ns |  1.653 ns |  2.030 ns |  83.44 ns |      - |         - |
|           CriteriaMatchingTerms | 833.25 ns | 15.723 ns | 14.708 ns | 832.17 ns | 0.0105 |      88 B |
|         CriteriaMatchingCreator |  98.58 ns |  1.999 ns |  2.380 ns |  97.45 ns |      - |         - |
|      CriteriaMatchingBeatmapIDs | 574.90 ns | 10.771 ns | 10.075 ns | 569.84 ns | 0.0105 |      88 B |
```

The benchmarks which filter with terms and execute `GetSearchableTerms` are significantly faster and allocate 10x less memory.